[EWT-585][SEDONA-406] Add as_numpy_masked method to retrieve bands data with nodata values masked as np.nan (#126)
* Add as_numpy_masked method to retrieve bands data with nodata value masked as np.nan

* [SEDONA-406] Make rasterio DatasetReader constructed from in-db raster has proper `crs` attribute (#127)

* in-db raster has a proper crs property

* Fix for old GDAL version (3.3.x)

+++ python/sedona/raster/sample_model.py
@@ -151,7 +151,7 @@ class SinglePixelPackedSampleModel(SampleModel):
                 for mask, bit_offset in zip(self.bit_masks, self.bit_offsets):
                     pixel.append((value & mask) >> bit_offset)
                 pixel_data.append(pixel)
-        arr = np.array(pixel_data).reshape(self.height, self.width, num_bands)
+        arr = np.array(pixel_data, dtype=bank_data.dtype).reshape(self.height, self.width, num_bands)
         return np.transpose(arr, [2, 0, 1])
 
 
@@ -188,6 +188,6 @@ class MultiPixelPackedSampleModel(SampleModel):
                 pos += 1
                 value = bank_data[pos]
                 shift = 0
-            band_data.append(np.array(pixels))
+            band_data.append(np.array(pixels, dtype=bank_data.dtype))
 
         return np.array(band_data).reshape(1, self.height, self.width)
